requirements.txt
# PDF Generation Agent - Requirements
# Python 3.8+

# Web framework
Flask>=2.3.0

# Markdown processing
Markdown>=3.4.0

# HTTP library for API requests
requests>=2.31.0

# Optional: For better path handling (included in Python 3.4+)
pathlib2>=2.3.7; python_version < '3.4'

# Optional: For environment variable management
python-dotenv>=1.0.0
